scripts/requirements.txt
click
pyyaml
unittest
